+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Fixed the target indicator invisible for non Epic Fight patched entities
+- Fixed the lock-on target not being synced to the server when using mouse snap to change the target
+
+### Changed
+
+- Now Lock-on automatically searches a new target if there is no currently focusing entity
+
 ## [20.14.1] - 2025-12-07
 
 ### Added